refactor(domains): move demo-product-images to domains/catalog (#232)
