+++ src/domains/catalog/demo-product-images.ts
